change oracle native contract file struct
