+++ smartcontract/service/native/native_service.go
@@ -41,7 +41,6 @@ var (
 	Contracts = map[common.Address]RegisterService{
 		genesis.OntContractAddress: RegisterOntContract,
 		genesis.OngContractAddress: RegisterOngContract,
-		genesis.OracleContractAddress: RegisterOracleContract,
 	}
 )
 
@@ -113,11 +112,4 @@ func RegisterOngContract(native *NativeService) {
 	native.Register("transfer", OngTransfer)
 	native.Register("approve", OngApprove)
 	native.Register("transferFrom", OngTransferFrom)
-}
-
-func RegisterOracleContract(native *NativeService) {
-	native.Register("createOracleRequest", CreateOracleRequest)
-	native.Register("setOracleOutcome", SetOracleOutcome)
-	native.Register("setOracleCronOutcome", SetOracleCronOutcome)
-	native.Register("changeCronView", ChangeCronView)
 }
